+++ practice/tnam/echo_server/echo_client.cpp
@@ -1,74 +1,87 @@
-#include <iostream>
-#include <cstdlib>
-#include <cstring>
-#include <unistd.h>
-#include <arpa/inet.h>
-#include <sys/socket.h>
+// echo client
+// 1. 소켓 생성
+// 2. 서버 주소 설정
+// 3. 서버와 연결
+// 4. 데이터 전송
+// 5. 데이터 수신
 
-const int PORT = 194;
-const int BUFFER_SIZE = 1024;
+#include <iostream>
+#include <sys/socket.h>
+#include <string>
+#include <netinet/in.h>
+#include <arpa/inet.h>
+#include <unistd.h>
 
 int main(int argc, char** argv)
 {
 	if (argc != 2)
 	{
-		std::cerr << "Invalid argument" << std::endl;
+		std::cerr << "invalid argument." << std::endl;
+		return EXIT_FAILURE;
 	}
 
-    // 소켓 생성
-    int clientSocket = socket(AF_INET, SOCK_STREAM, 0);
-    if (clientSocket == -1)
+	// 1. 소켓 생성
+	int clientSocket = socket(AF_INET, SOCK_STREAM, 0);
+	if (clientSocket == -1)
 	{
-        std::cerr << "Error creating socket." << std::endl;
-        return EXIT_FAILURE;
-    }
+		std::cerr << "error creating socket." << std::endl;
+		return EXIT_FAILURE;
+	}
 
-    // 서버 주소 설정
-    sockaddr_in serverAddr;
-    std::memset(&serverAddr, 0, sizeof(serverAddr));
-    serverAddr.sin_family = AF_INET;
-    serverAddr.sin_addr.s_addr = inet_addr(argv[1]);  // 서버의 IP 주소
-    serverAddr.sin_port = htons(PORT);
+	// 2. 서버 주소 설정
+	sockaddr_in server;
+	char* serverIP = argv[1];
+	const int serverPort = 194;
+	std::memset(&server, 0, sizeof(server));
+	server.sin_family = AF_INET;
+	server.sin_addr.s_addr = inet_addr(argv[1]);
+	server.sin_port = htons(serverPort);
 
-    // 서버에 연결
-    if (connect(clientSocket, reinterpret_cast<struct sockaddr*>(&serverAddr), sizeof(serverAddr)) == -1)
+	// 3. 서버와 연결
+	if (connect(clientSocket, reinterpret_cast<sockaddr *>(&server), sizeof(server)) == -1)
 	{
-        std::cerr << "Error connecting to server." << std::endl;
-        close(clientSocket);
-        return EXIT_FAILURE;
-    }
+		std::cerr << "error connecting server." << std::endl;
+		close(clientSocket);
+		return EXIT_FAILURE;
+	}
+	std::cout << "Connected server : " << serverIP << std::endl;
 
-    std::cout << "Connected to server." << std::endl;
-
-    // 에코 메시지 전송 및 수신
-    char message[BUFFER_SIZE];
-    while (true)
+	// 4. 데이터 전송
+	char sendBuffer[1000];
+	ssize_t sendLength = 0;
+	char recvBuffer[1000];
+	ssize_t recvLength = 0;
+	while (true)
 	{
-        std::cout << "Enter message (or 'exit' to quit): ";
-        std::cin.getline(message, sizeof(message));
-
-        if (std::strcmp(message, "exit") == 0) 
+		std::cout << "Enter the message(\"exit\" to quit): " << std::endl;
+		std::cin.getline(sendBuffer, sizeof(sendBuffer));
+		if (std::strcmp(sendBuffer, "exit") == 0)
 		{
-            break;
-        }
+			break;
+		}
 
-        // 메시지 전송
-        send(clientSocket, message, std::strlen(message), 0);
-
-        // 에코 메시지 수신
-        int bytesRead = recv(clientSocket, message, sizeof(message), 0);
-        if (bytesRead <= 0)
+		// Todo: 데이터 전송
+		sendLength = send(clientSocket, sendBuffer, sizeof(sendBuffer), 0);
+		if (sendLength <= 0)
 		{
-            std::cerr << "Error receiving data from server." << std::endl;
-            break;
-        }
+			std::cerr << "error sending data." << std::endl;
+			close(clientSocket);
+			return EXIT_FAILURE;
+		}
+		std::cout << "send data " << sendLength << " bytes." << std::endl;
 
-        message[bytesRead] = '\0';
-        std::cout << "Received from server: " << message << std::endl;
-    }
+		// Todo: 데이터 수신
+		recvLength = recv(clientSocket, recvBuffer, sizeof(recvBuffer), 0);
+		if (recvLength <= 0)
+		{
+			std::cerr << "error receiving data." << std::endl;
+			close(clientSocket);
+			return EXIT_FAILURE;
+		}
+		std::cout << "receive data " << recvLength << " bytes. : " << recvBuffer << std::endl;
+	}
+	
+	close(clientSocket);
 
-    // 소켓 닫기
-    close(clientSocket);
-
-    return EXIT_SUCCESS;
+	return 0;
 }
